Refactor results intercepts: split mocks into separate files and update imports
- Moved OPAL_FINES_RESULTS_REF_DATA_MOCK and related data from ResultsIntercept.mocks.ts to ResultsSearchIntercept.mocks.ts.
- Updated ResultsIntercepts.ts to import mocks from their new locations.
- Adjusted import paths in FinesAccEnfOverrideAdd.cy.ts and changeEnforcementOverride.cy.ts to reflect the new structure.
