+++ cypress/component/CommonIntercepts/referenceData/results/EnforcementOverrideResultsIntercept.mocks.ts
@@ -1,0 +1,124 @@
+import { IOpalFinesResultsRefData } from '@services/fines/opal-fines-service/interfaces/opal-fines-results-ref-data.interface';
+
+export const OPAL_FINES_ENF_OVERRIDE_RESULT_REF_DATA_MOCK: IOpalFinesResultsRefData = {
+  count: 13,
+  refData: [
+    {
+      result_id: 'ABDC',
+      result_title: 'Application made for Benefit Deductions',
+      result_title_cy: 'Cais am dynnu arian o fudd-daliadau',
+      active: true,
+      result_type: 'Result',
+      imposition_creditor: '',
+      imposition_allocation_order: null,
+    },
+    {
+      result_id: 'AEOC',
+      result_title: 'Attachment of Earnings Order - with Collection Order',
+      result_title_cy: 'Gorchymyn atafaelu enillion - cyfrif dirwy',
+      active: true,
+      result_type: 'Result',
+      imposition_creditor: '',
+      imposition_allocation_order: null,
+    },
+    {
+      result_id: 'BWTD',
+      result_title: 'Bail Warrant - dated',
+      result_title_cy: 'Gwarant Mechniaeth - Gyda dyddiad',
+      active: true,
+      result_type: 'Result',
+      imposition_creditor: '',
+      imposition_allocation_order: null,
+    },
+    {
+      result_id: 'BWTU',
+      result_title: 'Bail Warrant - undated',
+      result_title_cy: 'Gwarant Mechniaeth - Heb Ddyddiad',
+      active: true,
+      result_type: 'Result',
+      imposition_creditor: '',
+      imposition_allocation_order: null,
+    },
+    {
+      result_id: 'CLAMPO',
+      result_title: 'Clamping Order',
+      result_title_cy: null,
+      active: true,
+      result_type: 'Result',
+      imposition_creditor: '',
+      imposition_allocation_order: null,
+    },
+    {
+      result_id: 'CWN',
+      result_title: 'Committal Warrant Notice',
+      result_title_cy: 'Rhybudd o Warant Traddodi',
+      active: true,
+      result_type: 'Result',
+      imposition_creditor: '',
+      imposition_allocation_order: null,
+    },
+    {
+      result_id: 'DW',
+      result_title: 'Warrant of Control',
+      result_title_cy: 'Gwarant Atafaelu',
+      active: true,
+      result_type: 'Result',
+      imposition_creditor: '',
+      imposition_allocation_order: null,
+    },
+    {
+      result_id: 'MAN',
+      result_title: 'Manual Enforcement',
+      result_title_cy: null,
+      active: false,
+      result_type: 'Result',
+      imposition_creditor: '',
+      imposition_allocation_order: null,
+    },
+    {
+      result_id: 'NBWT',
+      result_title: 'No bail warrant',
+      result_title_cy: 'Gwarant dim mechniaeth',
+      active: true,
+      result_type: 'Result',
+      imposition_creditor: '',
+      imposition_allocation_order: null,
+    },
+    {
+      result_id: 'REGF',
+      result_title: 'Registration of Fine',
+      result_title_cy: 'Rhybudd Cofrestru',
+      active: true,
+      result_type: 'Result',
+      imposition_creditor: '',
+      imposition_allocation_order: null,
+    },
+    {
+      result_id: 'SUMA',
+      result_title: 'Enforcement Summons (Automatic enforcement only)',
+      result_title_cy: 'Gwŷs Gorfodi',
+      active: true,
+      result_type: 'Result',
+      imposition_creditor: '',
+      imposition_allocation_order: null,
+    },
+    {
+      result_id: 'TFOOUT',
+      result_title: 'Transfer of Fine Order to a Court in England or Wales',
+      result_title_cy: 'Trosglwyddo Gorchymyn Dirwy i Lys yng Nghymru neu Lloegr',
+      active: true,
+      result_type: 'Action',
+      imposition_creditor: '',
+      imposition_allocation_order: null,
+    },
+    {
+      result_id: 'FSN',
+      result_title: 'Further Steps Notice',
+      result_title_cy: 'Rhybudd o gamau pellach',
+      active: true,
+      result_type: 'Result',
+      imposition_creditor: '',
+      imposition_allocation_order: null,
+    },
+  ],
+};
