+++ cypress/component/CommonIntercepts/referenceData/results/ResultsIntercept.mocks.ts
@@ -1,698 +1,3 @@
-import { IOpalFinesResultRefData } from '@services/fines/opal-fines-service/interfaces/opal-fines-result-ref-data.interface';
-import { IOpalFinesResultsRefData } from '@services/fines/opal-fines-service/interfaces/opal-fines-results-ref-data.interface';
-
-export const OPAL_FINES_RESULTS_REF_DATA_MOCK: IOpalFinesResultsRefData = {
-  count: 8,
-  refData: [
-    {
-      result_id: 'FCC',
-      result_title: 'Criminal Courts Charge',
-      result_title_cy: "Ffi'r Llysoedd Troseddol",
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: 'CF',
-      imposition_allocation_order: 6,
-    },
-    {
-      result_id: 'FCOMP',
-      result_title: 'Compensation',
-      result_title_cy: 'Iawndal',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: 'Any',
-      imposition_allocation_order: 1,
-    },
-    {
-      result_id: 'FCOST',
-      result_title: 'Costs',
-      result_title_cy: 'Costau',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '!CPS',
-      imposition_allocation_order: 3,
-    },
-    {
-      result_id: 'FCPC',
-      result_title: 'Costs to Crown Prosecution Service',
-      result_title_cy: 'Costau i Wasanaeth Erlyn y Goron',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: 'CPS',
-      imposition_allocation_order: 3,
-    },
-    {
-      result_id: 'FFR',
-      result_title: 'FORFEITED RECOGNISANCE',
-      result_title_cy: null,
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: 'CF',
-      imposition_allocation_order: 10,
-    },
-    {
-      result_id: 'FO',
-      result_title: 'Fine',
-      result_title_cy: 'Dirwy',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: 'CF',
-      imposition_allocation_order: 5,
-    },
-    {
-      result_id: 'FVEBD',
-      result_title: 'Vehicle Excise Back Duty',
-      result_title_cy: 'Ôl-dreth ar y Dreth Cerbyd',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: 'CF',
-      imposition_allocation_order: 8,
-    },
-    {
-      result_id: 'FVS',
-      result_title: 'Victim Surcharge',
-      result_title_cy: 'Gordal Dioddefwr',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: 'CF',
-      imposition_allocation_order: 2,
-    },
-  ],
-};
-
-export const OPAL_FINES_RESULT_REF_DATA_MOCK: IOpalFinesResultRefData[] = [
-  {
-    result_id: 'REM',
-    result_title: 'Reminder of Unpaid Fine',
-    result_title_cy: 'Nodyn atgoffa terfynol am ddirwy heb ei thalu',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: false,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: false,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: false,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: false,
-    lists_monies: false,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false" }]',
-  },
-  {
-    result_id: 'FO',
-    result_title: 'Fine',
-    result_title_cy: 'Dirwy',
-    result_type: 'Result',
-    active: true,
-    imposition_allocation_priority: 6,
-    imposition_creditor: 'CF',
-    imposition: true,
-    imposition_category: 'Fines',
-    imposition_accruing: false,
-    enforcement: false,
-    enforcement_override: false,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: false,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: false,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: false,
-    lists_monies: false,
-    generates_warrant: false,
-    result_parameters: undefined,
-    allow_payment_terms: true,
-    requires_employment_data: false,
-    allow_additional_action: false,
-    enf_next_permitted_actions: undefined,
-    requires_lja: false,
-    manual_enforcement: false,
-  },
-  {
-    result_id: 'ABDC',
-    result_title: 'Application made for Benefit Deductions',
-    result_title_cy: 'Cais am dynnu arian o fudd-daliadau',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: true,
-    further_enforcement_disallow: false,
-    enforcement_hold: false,
-    requires_enforcer: true,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: false,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: true,
-    lists_monies: false,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false", "hint":"For example, LIVE UC or AMEND UC +400.00" },{ "name":"enforcer", "prompt":"Job Center Plus code (Enforcer code)", "type":"enforcers", "min":1, "max":1, "language_dependent":"true" }]',
-  },
-  {
-    result_id: 'AEOC',
-    result_title: 'Attachment of Earnings Order - with Collection Order',
-    result_title_cy: 'Gorchymyn atafaelu enillion - cyfrif dirwy',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: true,
-    further_enforcement_disallow: false,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: false,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: true,
-    lists_monies: false,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false", "hint":"For example, include employer name" }]',
-  },
-  {
-    result_id: 'BWTD',
-    result_title: 'Bail Warrant - dated',
-    result_title_cy: 'Gwarant Mechniaeth - Gyda dyddiad',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: true,
-    enforcement_hold: false,
-    requires_enforcer: true,
-    generates_hearing: true,
-    collection_order: false,
-    extend_ttp_disallow: true,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: true,
-    lists_monies: true,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false" },{ "name":"hearingdate", "prompt":"Adjourned to", "type":"date", "min":1, "language_dependent":false, "hint":"For example, 31/01/2023" },{ "name":"courtcode", "prompt":"Court code", "type":"integer", "min":1, "max":999, "language_dependent":false },{ "name":"enforcer", "prompt":"Warrant enforcer (Enforcer code)", "type":"enforcers", "min":1, "max":1, "language_dependent":"true" }]',
-  },
-  {
-    result_id: 'BWTU',
-    result_title: 'Bail Warrant - undated',
-    result_title_cy: 'Gwarant Mechniaeth - Heb Ddyddiad',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: true,
-    enforcement_hold: false,
-    requires_enforcer: true,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: true,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: true,
-    lists_monies: true,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false" },{ "name":"baildirection", "prompt":"Date and time (Bail direction)", "type":"text", "min":0, "max":500, "language_dependent":false, "hint":"As ordered in court. For example, 31/01/2023 and 14:30" },{ "name":"courtdetails", "prompt":"Court details", "type":"text", "min":1, "max":100, "language_dependent":false },{ "name":"enforcer", "prompt":"Enforcer code", "type":"enforcers", "min":1, "max":1, "language_dependent":"true" }]',
-  },
-  {
-    result_id: 'CLAMPO',
-    result_title: 'Clamping Order',
-    result_title_cy: '',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: true,
-    enforcement_hold: false,
-    requires_enforcer: true,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: true,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: true,
-    lists_monies: true,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false", "hint":"For example, vehicle details held on account" },{ "name":"effectivedate", "prompt":"Effective from date", "type":"date", "language_dependent":true, "hint":"For example, 31/01/2023" },{ "name":"enforcer", "prompt":"Clamping contractor code (Enforcer code)", "type":"enforcers", "min":1, "max":1, "language_dependent":"true" }]',
-  },
-  {
-    result_id: 'CWN',
-    result_title: 'Committal Warrant Notice',
-    result_title_cy: 'Rhybudd o Warant Traddodi',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: true,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: true,
-    collection_order: false,
-    extend_ttp_disallow: true,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: false,
-    lists_monies: true,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false", "hint":"For example, failed to comply with suspended committal" },{ "name":"hearingdate", "prompt":"Adjourned to", "type":"date", "min":1, "language_dependent":false, "hint":"For example, 31/01/2023" },{ "name":"courtcode", "prompt":"Court code", "type":"integer", "min":1, "max":999, "language_dependent":false }]',
-  },
-  {
-    result_id: 'DW',
-    result_title: 'Warrant of Control',
-    result_title_cy: 'Gwarant Atafaelu',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: true,
-    enforcement_hold: false,
-    requires_enforcer: true,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: true,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: true,
-    lists_monies: true,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    generates_warrant: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false" },{ "name":"enforcer", "prompt":"Enforcer code", "type":"enforcers", "min":1, "max":1, "language_dependent":"true" }]',
-  },
-  {
-    result_id: 'MAN',
-    result_title: 'Manual Enforcement',
-    result_title_cy: '',
-    result_type: 'Result',
-    active: false,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: false,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: false,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: false,
-    lists_monies: false,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters: undefined,
-  },
-  {
-    result_id: 'NBWT',
-    result_title: 'No bail warrant',
-    result_title_cy: 'Gwarant dim mechniaeth',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: true,
-    enforcement_hold: false,
-    requires_enforcer: true,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: true,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: true,
-    lists_monies: true,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false" },{ "name":"courtdetails", "prompt":"Court details", "type":"text", "min":0, "max":100, "language_dependent":false },{ "name":"enforcer", "prompt":"Enforcer code", "type":"enforcers", "min":1, "max":1, "language_dependent":"true" }]',
-  },
-  {
-    result_id: 'REGF',
-    result_title: 'Registration of Fine',
-    result_title_cy: 'Rhybudd Cofrestru',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: true,
-    further_enforcement_disallow: false,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: false,
-    extend_ttp_preserve_last_enf: true,
-    prevent_payment_card: false,
-    lists_monies: true,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false" }]',
-  },
-  {
-    result_id: 'SUMA',
-    result_title: 'Enforcement Summons (Automatic enforcement only)',
-    result_title_cy: 'Gwŷs Gorfodi',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: true,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: true,
-    collection_order: false,
-    extend_ttp_disallow: true,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: false,
-    lists_monies: true,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false" }]',
-  },
-  {
-    result_id: 'TFOOUT',
-    result_title: 'Transfer of Fine Order to a Court in England or Wales',
-    result_title_cy: 'Trosglwyddo Gorchymyn Dirwy i Lys yng Nghymru neu Lloegr',
-    result_type: 'Action',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: false,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: false,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: true,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: true,
-    lists_monies: false,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: true,
-    manual_enforcement: undefined,
-    result_parameters: undefined,
-  },
-  {
-    result_id: 'FSN',
-    result_title: 'Further Steps Notice',
-    result_title_cy: 'Rhybudd o gamau pellach',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: false,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: false,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: false,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: false,
-    lists_monies: false,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false" }]',
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    generates_warrant: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-  },
-];
-
-export const OPAL_FINES_ENF_OVERRIDE_RESULT_REF_DATA_MOCK: IOpalFinesResultsRefData = {
-  count: 13,
-  refData: [
-    {
-      result_id: 'ABDC',
-      result_title: 'Application made for Benefit Deductions',
-      result_title_cy: 'Cais am dynnu arian o fudd-daliadau',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'AEOC',
-      result_title: 'Attachment of Earnings Order - with Collection Order',
-      result_title_cy: 'Gorchymyn atafaelu enillion - cyfrif dirwy',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'BWTD',
-      result_title: 'Bail Warrant - dated',
-      result_title_cy: 'Gwarant Mechniaeth - Gyda dyddiad',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'BWTU',
-      result_title: 'Bail Warrant - undated',
-      result_title_cy: 'Gwarant Mechniaeth - Heb Ddyddiad',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'CLAMPO',
-      result_title: 'Clamping Order',
-      result_title_cy: null,
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'CWN',
-      result_title: 'Committal Warrant Notice',
-      result_title_cy: 'Rhybudd o Warant Traddodi',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'DW',
-      result_title: 'Warrant of Control',
-      result_title_cy: 'Gwarant Atafaelu',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'MAN',
-      result_title: 'Manual Enforcement',
-      result_title_cy: null,
-      active: false,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'NBWT',
-      result_title: 'No bail warrant',
-      result_title_cy: 'Gwarant dim mechniaeth',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'REGF',
-      result_title: 'Registration of Fine',
-      result_title_cy: 'Rhybudd Cofrestru',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'SUMA',
-      result_title: 'Enforcement Summons (Automatic enforcement only)',
-      result_title_cy: 'Gwŷs Gorfodi',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'TFOOUT',
-      result_title: 'Transfer of Fine Order to a Court in England or Wales',
-      result_title_cy: 'Trosglwyddo Gorchymyn Dirwy i Lys yng Nghymru neu Lloegr',
-      active: true,
-      result_type: 'Action',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'FSN',
-      result_title: 'Further Steps Notice',
-      result_title_cy: 'Rhybudd o gamau pellach',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-  ],
-};
+export * from './EnforcementOverrideResultsIntercept.mocks';
+export * from './ResultByCodeIntercept.mocks';
+export * from './ResultsSearchIntercept.mocks';
